Refactor upgrade effects parsing and application
